zq update through "Recruiter API in zqd for server and client " by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1695, authored by @marktwallace,
has been merged.

Recruiter API in zqd for server and client

The zqd recruiter manages a pool of zqd workers and provides a mechanism for zqd root processes to recruit zqd worker processes for query execution. This PR includes the server API provided by the recruiter process, and the client side of the same API that is used by the zqd worker processes.

The zqd recruiter is started with: zqd --personality=recruiter

The API is described in ppl/zqd/recruiter/README