zq update through "fix pcap space time range regression" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1798, authored by @alfred-landrum,
has been merged.

fix pcap space time range regression

This fixes a bug that I introduced in https://github.com/brimsec/zq/pull/1693 , when moving code to a new location.

I didn't notice this until it became apparent in the UI when importing a pcap. I'll add a test, but I verified this fixes https://github.com/brimsec/zq/issues/1797 .