zq update through "copy columns when looking up TypeRecord" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1811, authored by @mccanne,
has been merged.

copy columns when looking up TypeRecord

This commit makes a copy of the zng.Columns when looking up
a TypeRecord so that the caller mutate its underlying zng.Column
slice without harming the type context.