zq update through "fix bug with rows larger than max read buffer" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1808, authored by @mccanne,
has been merged.

fix bug with rows larger than max read buffer

This commit adds a better error message to the zng reader with the
length of a value exceeds the max read size.   We also added flags
to set the max read buffer (and target read buffer) sizes.

Closes brimsec/zq#1803