zq update through "add MarshalCustomRecord" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1819, authored by @mccanne,
has been merged.

add MarshalCustomRecord

This commit adds a function to the zng marshaler to create a record
from a slice of strings containing the field names and a corresponding
slice of values as []interface{}.  This lets a client of the marshaler
create a record with a reliable column order and specify the field names
dynamically instead of with fixed struct tags.