zq update through "allow outer records to be aliased" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1829, authored by @mccanne,
has been merged.

allow outer records to be aliased

This commit is a first step toward brimsec/zq#1801, which is to embed
zng.Value into zng.Record so that a record can take on any
alias type.  This step toward this goal adds an extra zng.Type
to record called Alias, which is either a zng.TypeRecord
(behavior as before) or a zng.TypeAlias (new behavior).

This allows the example ZSONs in the ZSON spec, which have
top-level type defs, to actually work.  Tests were adding
to cover these examples and make sure the typeof(.) operator
worked on the aliased outer records.