zq update through "Fill in (t *TypeTable).LookupType" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1832, authored by @henridf,
has been merged.

Fill in (t *TypeTable).LookupType

This method can't yet be exercised via the contact surface (though that should change soon with brimsec/zq#1785 and brimsec/zq#1619).

So I couldn't test it with a ztest. Instead, I added a temporary Go test in aab7b99, and then reverted it after it passed CI.