zq update through "wrap zng marshalers with state" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1831, authored by @mccanne,
has been merged.

wrap zng marshalers with state

This commit refactors the zng marshaling code to use state data structure
instead of passing it in.  This represents the fundamental notion that
a marshaler must have a type context as outside data is marshaled into
a zng stream (unlike JSON where there is no type state).

The MarshalContext struct embeds and exports zcode.Buidler and
resolver.Context so that custom marshalers may operate directly
on both sets of methods when build marshaled results and return
the proper (and newly constructed) zng.Types.

This also sets up the code for a subsequent PR that will add the
ability to unmarshal into an interface type (unlike json.Unmarshal),
where we will use the ZSON type name to determine what concrete
implementation should be unmarshaled into for a given interface value.