zq update through "json types: add support for inferring type of "extra" fields" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1842, authored by @henridf,
has been merged.

json types: add support for inferring type of "extra" fields

A new boolean field `inferextra`  is added to the json type config. If true, the json typed reader no longer returns an error upon encountering fields not present in the current descriptor, but instead applies plain old json type inference and adds them to the end of the record. If false, the reader follows the prior behavior of returning an error.

Fields are added to the end of the record, at the top level, with a "flattened" field name. For example, assuming the incoming json object has a top-level field `uid2` and a nested field `id2` that are not in the schema, such as:

```
...
{
...
"uid2": "asdfasdfasdf",
"id2": {"orig_h": "2.2.2.1"}
}
```
then those fields will show up as "uid" and "id2.orig_h", both at the record top-level. On the one hand, this is is not very ergonomic, but on the other, it is necessary to allow "pushing" the extra fields to the right. And perhaps more importantly, leaving the inferred fields "in their place", would require a substantially bigger change to the typed parser, which does not seem worthwhile given its expected TTL.

First part of brimsec/zq#1841