zq update through "create zson.Formatter from most of zsonio.Writer" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1840, authored by @mccanne,
has been merged.

create zson.Formatter from most of zsonio.Writer

This commit creates zson.Formatter from the logic that used to
live in zsonio.Writer.  With this refactor, a coming-soon
zson.Marshaler can use the formatter without creating an
import cycles.

This is mostly code movement and rename.  The one logic change
is the formatter now uses a stack of strings.Builder instead of
saving the writer pointer in the stack when re-entering.  Also,
instead of writing the formatted output to an io.Writer, the
output is built in memory using a strings.Bulder.

Closes brimsec/zq#1763