zq update through "Add typing config for Suricata "vlan" field" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1851, authored by @philrz,
has been merged.

Add typing config for Suricata "vlan" field

A community user experienced the symptom that led to brimsec/zq#1841 and was able to point to a publicly-available pcap that reproduces the issue (https://download.netresec.com/pcap/maccdc-2012/maccdc2012_00002.pcap.gz). With this I was able to confirm that `vlan` is a top-level field in Suricata alerts that the JSON typing config in GA `zq` tagged `v0.25.0` was not prepared to accept. An example alert generated from that pcap:

```
{
  "timestamp": "2012-03-16T07:22:31.400000-0700",
  "flow_id": 444124526741776,
  "pcap_cnt": 1083767,
  "event_type": "alert",
  "vlan": [
    120
  ],
  "src_ip": "192.168.24.252",
  "src_port": 80,
  "dest_ip": "192.168.202.101",
  "dest_port": 49147,
  "proto": "TCP",
  "community_id": "1:Gd8IfCRv2AjAf3KvCReRBdLZepk=",
  "tx_id": 0,
  "alert": {
    "action": "allowed",
    "gid": 1,
    "signature_id": 2221021,
    "rev": 1,
    "signature": "SURICATA HTTP response header invalid",
    "category": "Generic Protocol Command Decode",
    "severity": 3
  },
  "app_proto": "http"
}
```

This PR includes the typing config that covers this. Here's an example of Brim happily now accepting the field at import.

![image](https://user-images.githubusercontent.com/5934157/102733248-963c0700-42f1-11eb-9a41-bf6e33d8d8ff.png)