zq update through "Simplified Recruiter API" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1826, authored by @marktwallace,
has been merged.

Simplified Recruiter API

This is a substantial change to the recruiter design from last week. Read the new design doc: ppl/zqd/recruiter/README.md to understand the select's and for loops in ppl/zqd/recruiter/workerreg.go.

https://github.com/brimsec/zq/blob/recruiter-alt/ppl/zqd/recruiter/README.md

The following google doc describes what changed in the design:
https://docs.google.com/document/d/1Y1mM7bpkBnTvwnMPFNdACNu9ELAPInYtFQhBf1d-C1Y/edit?usp=sharing