zq update through "add zson marshaler and clean up zng marshaler" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1854, authored by @mccanne,
has been merged.

add zson marshaler and clean up zng marshaler

This commit adds a zson marshaler, which is basically a wrapper
around zng marshal that allows one to do marshaling strictly
with ZSON text format and not worry about zng or package zng/resolver.

This work also exposed some problems with the zng marshaler regarding
interfaces so several bug fixes there included here as well as
an extension of the zng marshaler to handle empty interface for
non-struct types.  We also made things work when records have
recursively invested interface types (which are resolvable with
marshaler bindings).