zq update through "Don't pass in flowgraph warnings channel from outside" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1865, authored by @henridf,
has been merged.

Don't pass in flowgraph warnings channel from outside

This seems like an improvement, prompted by getting confused by the client-supplied warnings channel while working on brimsec/zq#1833 (I initially thought that the point of passing it in was that I could then read from it... but of course the driver reads from it). From the commit:

Previously, the driver package allowed clients to pass in a flowgraph
warnings channel. This was used by clients that created a channel for
use by a zbuf.WarningReader, and then passed that channel to a
flowgraph.

This commit changes zbuf.WarningReader to take a callback instead. It is
a slightly simpler API and when used in conjunction with the driver package,
the pattern is cleaner: the driver creates its warning channel
internally, and the neither WarningReader nor the client plumbing
things together needs to know about either one.