zq update through "Fix flaky fs.TailFile related tests" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1877, authored by @mattnibs,
has been merged.

Fix flaky fs.TailFile related tests

We've been experiencing some test failures on master due to the fact that
windows [1] (and sometimes linux [2]) appears to need an extra sync call in
order to trigger a "write" event on the file system event watcher.

Add some extra Sync calls in tests where needed and reformat the
TestTailFileWrite test to ensure its actually testing that write events are
triggered.

As also TailFileWithContext variant so the TestTailFileWrite tests fails sooner
should it not receive "write" events.

[1] https://github.com/brimsec/zq/runs/1601524036?check_suite_focus=true
[2] https://github.com/brimsec/zq/runs/1597423491?check_suite_focus=true