zq update through "Ingest shaper" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1870, authored by @henridf,
has been merged.

Ingest shaper

This commit adds support for a running Z "shaper" script on incoming logs. This facility is intended for ETL tasks such as transforming record layouts and field names, casting leaf types, or policing incoming types against a set of admissible schemas. (Z script does not quite yet have the support for much of this, but that is imminent).

Closes brimsec/zq#1833