zq update through "move compiler entry points from package zql to package compiler" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1923, authored by @mccanne,
has been merged.

move compiler entry points from package zql to package compiler

This commit moves the entry points for compilation for the zql package
to the compiler package.  This will allow a future PR to put semantic
analysis behind the compiler entry points in a uniform way from within
the compiler package