zq update through "remove unused data-flow analysis from the optimizer" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1916, authored by @mccanne,
has been merged.

remove unused data-flow analysis from the optimizer

This commit removes the unused data-flow analysis that is in the
compiler optimizer, namely computeColumns(), colset, and supporting
functions.  The current logic that parallelizes the flowgraph does not
rely upon this analysis.  We have spent much collective time maintaining
this code and the related tests as the flowgraph and expr/proc logic
has evolved.  We will want such analysis down the road (e.g., to support
predicate-pushdown in the column reader) and at that point,
the requirements will likely be different as the AST and flowgraph
capabilities mature, and we will likely want to use a different
design pattern based on use-def dataflow analysis from the compiler
literature.