zq update through "iosrc.URI: Support parsing, encoding UNC paths" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1929, authored by @mattnibs,
has been merged.

iosrc.URI: Support parsing, encoding UNC paths

UNC paths can include support for a network based file path. Adjust
iosrc.ParseURI to keep hostname of a UNC path with a host. If a file uri has
a host, output the path as a UNC path.

Closes brimsec/zq#1927