zq update through "EKS smoke test" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1910, authored by @marktwallace,
has been merged.

EKS smoke test

This automates a very simple set of smoke tests to check if a zqd image can be deployed on EKS and works for basic distributed queries. An important feature of the smoke test is that it does not reserve exclusive resources, so it can run in parallel with itself.

The tests are simple, but the overall process is non-trivial. Since there is a lot to unpack here, I am interspersing comments in my own PR.

This adds two new jobs to what we have been running for PR's. The new jobs make extensive use of S3, so they will fail in ways that minio based testing does not.

The manual workflow dispatch dialog will be available only after this is pushed to master.