zq update through "fix flattener for nested nulls" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1954, authored by @mccanne,
has been merged.

fix flattener for nested nulls

This commit fixes a bug in the flattener where nested nulls were
not properly output for every column in the flattened record
(i.e., a null value for an embedded record did not properly recurse
into the nested record so only one null value was output instead
of a null for each flattened value).

While fixing this we added zcode.AppendNull() to clarify that the API
for appending null is the same for primitives and containers.

There are also some type context problems I noticed in fuse and the
csv writer but these can be fixed later since the code happens to work
as is (see brimsec/zq#1953)

Fixes brimsec/zq#1942