zq update through "make fuse stable" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1958, authored by @mccanne,
has been merged.

make fuse stable

This commit changes the fuser so the output type columns of
the union schema are stable.  Previously, the column order
was dependent on the type IDs, which could be non-deterministic
when scanning parallel inputs.  The new code determines the
column order from the order of the types encountered in the
incoming stream.

This fix will address intermittent CI failures due to this indeterminacy.

Fixes brimsec/zq#1702