zq update through "expr.LookupValueCaster(): change argument type and rename " by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1951, authored by @henridf,
has been merged.

expr.LookupValueCaster(): change argument type and rename

This commit changes `expr.LookupValueCaster()` to take a zng.Type rather than the string representation of the type. This is needed so that the forthcoming `cast()` function (now mostly working on branch shaping-function-cast) can call with the various leaf zng.Type's found during record traversal.

Also, while I was making this change, I realized ValueCaster is misleading and changed the name to PrimitiveCaster.