zq update through "zqd listen: Reformat startup logs" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1957, authored by @mattnibs,
has been merged.

zqd listen: Reformat startup logs

Create a named logger for manager, db and core and log when each of these
services has started, along with configuration.

Most important here is the "Connected" log line for the db logger, which prints out a password/username redacted version of the connection url used.

Closes brimsec/zq#1947