zq update through "remove redundant ordered merging during archive pcap ingest" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1973, authored by @alfred-landrum,
has been merged.

remove redundant ordered merging during archive pcap ingest

We sort data during the Write calls to archive.Writer, so an ordered
merge beforehand is unnecessary. Additionally, since the underlying
streams of records from the zeek and suricata analyzers are not
necessarily in order, it can cause records to not appear progressively
during pcap ingest (brimsec/zq#1971).