zq update through "Replace zbuf.ReadBatch with zbuf.NewPuller" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1962, authored by @nwt,
has been merged.

Replace zbuf.ReadBatch with zbuf.NewPuller

zbuf.ReadBatch can signal EOS in two ways: by returning a nil batch and
nil error or a short batch and nil error.  This undocumented behavior
leads to a bug in zbuf.Scanner.Pull, which does not check for a short
batch.  Fix the bug by replacing ReadBatch with zbuf.NewPuller.

Alternative to brimsec/zq#1894 (from which `ppl/zqd/ztests/zapi-postpath-errors.yaml` is copied).