zq update through "allow making file store spaces read only" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1952, authored by @alfred-landrum,
has been merged.

allow making file store spaces read only

This change adds a temporary variable in the api package that determines
whether file storage spaces are read only (and may not be created), and fixes
issues that arise when it is set. This also adds a hidden flag to zqd
listen to control it.