zq update through "archivestore: Delay before purge" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1983, authored by @mattnibs,
has been merged.

archivestore: Delay before purge

When running a compaction on an archivestore, after the Compact phase is
complete, add a one minute delay before a Purge run. This
is meant as a stop-gap solution for the bug encountered in brimsec/zq#1960.

Closes brimsec/zq#1960