zq update through "Notify slack when master workflow fails" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1986, authored by @mattnibs,
has been merged.

Notify slack when master workflow fails

Add 'Notify master failure' github action which is triggered whenever
a workflow_run on branch master is completed. If the workflow failed the action
will notify the BrimHQ slack of the failure.

Closes brimsec/zq#1963