zq update through "Listen on 127.0.0.1 in pkg/httpd tests" by nwt
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1992, authored by @nwt,
has been merged.

Listen on 127.0.0.1 in pkg/httpd tests

On macOS, if the firewall is enabled, running the pkg/httpd tests
triggers a "Do you want the application “httpd.test” to accept incoming
network connections?" dialog because the tests listen on 0.0.0.0.
Stifle the dialog by listening on 127.0.0.1 instead.