zq update through "Perf update at 806aadb, now including ZSON" by philrz
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1974, authored by @philrz,
has been merged.

Perf update at 806aadb, now including ZSON

Now that ZSON is available as part of the zq-sample-data (https://github.com/brimsec/zq-sample-data/pull/18) and we fixed the recently-found bug that was causing ZSON reads to fail (brimsec/zq#1966) we can include ZSON in the permutations for periodic snapshotting of zq's performance.

A brief summary of what I see in this update:

* The numbers have generally improved across the boards.
* Perhaps unsurprisingly, in the area of text formats, ZSON seems to perform a bit better than NDJSON, and worse than TZNG.