zq update through "Master chart for helm install with Postgres" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/1976, authored by @marktwallace,
has been merged.

Master chart for helm install with Postgres

Reworks the existing helm charts, converting them to subcharts of an umbrella chart called zservice. Simplifies and streamlines helm install and uninstall. This makes it easier to understand the installation and will allow us to more quickly add additional components including the API Server and Redis.

This changes the naming conventions of the deployed services. Instead of having a postfix of zqd, they now have a prefix of z. E.g:

recruiter-zdq --> z-recruiter