zq update through "initial multitenancy support" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2015, authored by @alfred-landrum,
has been merged.

initial multitenancy support

This changes adds initial multitenancy support to zqd.

The high level summary is that when requests are issued to zqd, those requests will have an associated identity, either specified by an authenticated JWT, or the "anonymous" identity. When a space is created, the tenant id of the creating identity is stored with it. Only users with the same tenant id will be allowed to interact with the space. When a list of spaces is requested, the tenant id is used as a filter in the returned spaces.

The token validation and context interaction parts of auth moved into a separate package from package zqd & api. I dropped the dependency on the auth0 jwt middleware, as it was just a thin wrapper around the go-jwt 'jwt.Parse' call. The validation code is better tested here with the new Go tests, and I wrote those guided by looking at test coverage results in my IDE.

zapi gets a new flag, `-configpath`, used to set where it looks for its configuration holding directory. It defaults to os.UserConfigDir, which is $HOME/.config for Linux systems. To aid this change, I made sure that the $HOME variable is set during ztests; it's set to the temporary directory where the ztest is run.

With this change, it's always possible to call `auth.IdentityFromContext`. When zqd is run without authentication enabled, it will always return an identity with the anonymous user and tenant values.

I added migrations to the file and postgres database implementations, adding a column for "tenant_id", which is set to the anonymous tenant value by default.

I added ztests that verify the tenant handling when zqd is backed by either a filedb or postgres. The ztests uses a new utility, "testtoken", to generate a signed JWT using the public/private test key pair in zqd/testdata.