zq update through "print any log files when awaitfile fails" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2023, authored by @alfred-landrum,
has been merged.

print any log files when awaitfile fails

When `awaitfile` fails, typically because of a failure of a zqd instance to start, make sure to output any .log files found. This will capture the log files for the zqd root, recruiter, and worker instances in some of the ztests that verify distributed queries.