zq update through "Update deploy job to use ZTest" by marktwallace
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2004, authored by @marktwallace,
has been merged.

Update deploy job to use ZTest

This adds a new set of ZTests tagged with "cluster" to run the same tests that were previously run directly with zapi in cd.yaml.

A Makefile target, test-cluster, that runs the cluster tests has been introduced. The cluster tests can be more easily extended and maintained as ZTests.

In addition, it addresses two existing intermittent failures in the Deploy test in cd.yaml.
(1) On some test failures, the temporary namespace (e.g. ci-408) was not being removed after the test. This PR consolidates three previous jobs: pre-clean, eks-test-deploy, and eks-test-zapi-1, into a single job that shares the same cleanup code that removes the temporary namespace.
(2) There was a "sleep 2" after the port-forward commands. It is replaced by two loops that verify the port-forward with netcat before proceeding.