zq update through "refactor package filter into expr and string-search components" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2026, authored by @mccanne,
has been merged.

refactor package filter into expr and string-search components

This commit refactors the filter package into the pieces that
perform the runtime logic (moved to package expr) and the pieces
that do generic string scanning (moved to pkg/stringsearch).

filter.Predicate has been renamed expr.Boolean and filter.Filter
is now expr.Filter.  expr.Boolean applies a predicate to a
zng.Value while expr.Filter applies a predicate to a zng.Record.

These changes are strictly code movement and renaming.  There are
no semenantic changes to the code here.

This refactoring is needed by the upcoming PR that unifies the search
and expression syntax and machinery.