zq update through "archive: Immutable local cache" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2020, authored by @mattnibs,
has been merged.

archive: Immutable local cache

Add ability to enable local caching of small immutable files (currently this is
only chunk metadata) in an archive. Add prometheus metrics for tracking cache
hits and misses on gets for small files.

Closes brimsec/zq#1898