zq update through "finish up zson scanner to read input incrementally" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2031, authored by @mccanne,
has been merged.

finish up zson scanner to read input incrementally

This commit addresses the placeholder logic in the zson scanner
that reads in the entire input upon open to simplify the scanning
logic.  Now that the zson parser is debugged and complete, this
change fixes the lexer to incrementally read input from the io.Reader
as needed.

Fixes brimsec/zq#1802