zq update through "tighten whitespace in zson -pretty=0 output" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2030, authored by @mccanne,
has been merged.

tighten whitespace in zson -pretty=0 output

This commit changes the zson formatter to drop the space after
a field colon when pretty printing is disabled.  The space between
a value and a type decorator is retained, but otherwise, all
spaces should be gone now for line-delimited zson output.

Closes brimsec/zq#1993