zq update through "Change rename to be a proc.Function" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2049, authored by @henridf,
has been merged.

Change rename to be a proc.Function

This is motivated by brimsec/zq#1959, which will use the newly-exposed rename function to rename conflicting fields.

(currently stacked on top of @mccanne 's rename fix branch; will change base to master after that one merges).