zq update through "Windows zq: propery route - path to stdin" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2061, authored by @mattnibs,
has been merged.

Windows zq: propery route - path to stdin

The current logic in zq was translating the stdin path arg '-' in a zq command
to /dev/stdin. While this works in unix environments this does not work for
windows. Instead translate '-' to the iosrc.Stdin representation
'stdio://stdin' so the path is properly routed to stdin in windows.

Closes brimsec/zq#2000