zq update through "rename archive package to lake" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2053, authored by @alfred-landrum,
has been merged.

rename archive package to lake

This change renames the archive package to lake, the `Archive` struct to `Lake`, and the canonical variable name for a `*Archive` from `ark` to `lk`. No other changes (api routes, metric names, etc) are changed.