zq update through "fix scanning of zson primitives" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2058, authored by @mccanne,
has been merged.

fix scanning of zson primitives

The logic to scan zson primitive values did not properly handle the
recent changes to read input as a stream and had presumed all of
the needed input was in the cursor.  This commit fixes peekPrimitive()
to make sure there is data up to the next whitespace, comma, or EOF
(since none of this can appear in a primitive value).

I would spend a bit more time here to do a bit more review and more
comprehensive tests at this point, but there are other priorities so I
created issue brimsec/zq#2057

Fixes brimsec/zq#2056