zq update through "handle null casts properly for all Z types" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2077, authored by @mccanne,
has been merged.

handle null casts properly for all Z types

The cast expressions are still a work in progress, but in particular,
the various casters aren't handling null properly.  Since Z allows
any value to be null, this is very easy handed as universal check
before calling the type-specific caster.  This commit adds this logic
and a simple test to cover it.

This is related to issue brimsec/zq#1603 and fixes brimsec/zq#2065