zq update through "add join/split string functions" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2098, authored by @mccanne,
has been merged.

add join/split string functions

This commit adds functions for join() and split() with respect to strings.
Since Z is dynamically typed, this functions may do more down the road
based on the types of their arguments, but for now they are just string
functions.

Closes brimsec/zq#2089, brimsec/zq#2090