zq update through "make connection.Search easier to use" by alfred-landrum
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2105, authored by @alfred-landrum,
has been merged.

make connection.Search easier to use

connection.Search is a convenience wrapper around connection.SearchRaw. This change makes it a little easier to use, and adds example usage. It roundtrips with the server to parse the search query, to be more resilient against AST changes until we add support for AST versioning.