zq update through "fix join bug with memory management" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2108, authored by @mccanne,
has been merged.

fix join bug with memory management

The join proc did not correctly do a zng.Record.Keep() while building
the join set array causing the underlying batch to be freed while
record references still existed.  The fix is to call keep in the right
place.  Also, the joinkey memory could also be freed while in use
so it's now copied to a buffer in the join proc struct.

This bug happened more often with a 'join' proc at the front of the
pipeline (with no split) because the splitter proc otherwise keeps a ref on each
batch so this bug was better hidden from the downstream consumer.