zq update through "Fix ztest" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2155, authored by @henridf,
has been merged.

Fix ztest

Due to a variable sharing bug, ztest was re-running the same test under the guise of all the different tests it was running.

The first commit in this PR illustrates that with an invalid test that passes.
The second commit fixes ztest which fails the bad test. The third commit removes. the bad test.