zq update through "Fuse aggregator" by henridf
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2115, authored by @henridf,
has been merged.

Fuse aggregator

This PR adds the `fuse()` aggregator (with a placeholder name per https://github.com/brimsec/zq/issues/2114).

Upon encountering collisions (columns with same name but different types), it computes the type union of the types rather than the `_n` suffixes that are used by the fuse proc.

With this it should be straightforward to add "null absorption" (@mccanne 's "nullify"); we'll have to decide if that should be the default or something else.

close brimsec/zq#1965