zq update through "fix bug with match() and boolean ops on funcs in search context" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2152, authored by @mccanne,
has been merged.

fix bug with match() and boolean ops on funcs in search context

This commit fixes a bug where parsing of function-expressions
in proc context wasn't quite working right.  This change cleans
up and simplifies the grammar guards.

Fixes brimsec/zq#2150