zq update through "clean up conventions for missing values" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2154, authored by @mccanne,
has been merged.

clean up conventions for missing values

This commit creates zng.ErrMissing and zng.Missing to work toward
a unified way of handling missing values (i.e., expressions that
refer to fields that don't exist).  ErrMissing is a Go error
that can be returned up the chain to signal this condition while
Missing is a value that can be returned in zng data when a
result calls for embedding a Missing value instead of ignoring
it all together.

Closes brimsec/zq#2130