zq update through "make implied-type zson more user friendly" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2160, authored by @mccanne,
has been merged.

make implied-type zson more user friendly

This commit modifies the zson formatter to drop all type decorators
when a complex value is fully implied.  Null values are still
decorated and when their type is implied but complex, an embedded
typedef appears making subsequent nulls of the same type less
busy in the output.

Closes brimsec/zq#2118