zq update through "break out compiler package into sub-packages" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2170, authored by @mccanne,
has been merged.

break out compiler package into sub-packages

This commit refactors the compiler package so that it is now
comprised of semantic analysis (compiler/semantic) and runtime
generation of the semantic output to the Z runtime kernel
represenation (aka the proc.Interface).

No new code or logic is introduced here.  A subsequent PR will use
this refactoring to create a new flowgraph-based DSL that will
server as the intermediate form between the parser's AST and
the Z kernel runtime flowgraph.