zq update through "change -z to -I and allow multiple files" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2180, authored by @mccanne,
has been merged.

change -z to -I and allow multiple files

This commit changes the flag (-z) for running zq with the Z query in
a file.  The flag is replaced with -I (for "include") and the behavior
is a bit different: -I can be used multiple times and it is combined
with the normal Z query on the command line.  These combinations can
create some confusing syntax errors so the Z text parsed is display
when a syntax error occurs.

We also changed the "ZQL" name convention to "Z" in the command help.