zq update through "lake: Parallelize tsdir fetch" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2162, authored by @mattnibs,
has been merged.

lake: Parallelize tsdir fetch

The walk of lake stored in s3 is slowed greated in that tsDirs are fetched
sequentially. The main thread sits by idly while the metadata for all chunks in
a tsdir are retrieved over the network. Speed this process up by running the
tsdir fetches each in a separate goroutine.

Because it is important that results be returned in order, save the retrieved
tsdirs in a min/max heap, and release results from the stream when the next
min/max is received.

Closes brimsec/zq#2017