zq update through "add taste operator" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2200, authored by @mccanne,
has been merged.

add taste operator

This commit adds a "taste" operator (aka proc) that provides
a taste of the shape of the data.  It takes an optional
field-expression argument and computes "taste=any(e) by shape=typeof(e)"
where e is "." if it is omitted.  This is a nice shortcut
for demo, docs, and articles.

There is actually no new proc in the runtime as the parser
does a direct translation to the group-by proc.