zq update through "deprecate -t from output flags and add -z and -Z" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2208, authored by @mccanne,
has been merged.

deprecate -t from output flags and add -z and -Z

This commit removes the -t shortcut from the output flags
of zq etc as we are demoting the status of the tzng format
in favor of zson.  To this end, we added -z and -Z.
-z is a shortcut for "-f zson -pretty=0" providing line
oriented zson while -Z is a shortcut for "-f zson", providing
jq-like formatted printy-printing of zson.

We also removed the pre-existing "-z" option for "zar find"
in favor of picking the output format with "-f" just as Al
had suggested today that all tools adopt this convention.