zq update through "fix bug where random data looked like non-record zng value" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2206, authored by @mccanne,
has been merged.

fix bug where random data looked like non-record zng value

This commit fixes a bug where bad data that happens to parse
as zng caused a panic when the type ID of the zng value was
was not a record type.  The fix is to check the that top-level
values are records and return an error if they don't, which
otherwise allows the auto-detector to complete its task
and report an error.

Fixes brimsec/zq#2205