zq update through "Fix groupby sorted input nulls order" by mattnibs
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2220, authored by @mattnibs,
has been merged.

Fix groupby sorted input nulls order

If a sorted input is in descending order set nullsMax for the compare function
to false. This follows the behavior of the sort proc on sorted inputs where
null values are always printed last.

This fixes a bug when performing a "count every [duration]" that sporadically
caused records with nil time values to get emitted out-of-order.

Closes brimsec/zq#1912