zq update through "simplify ast unpacker" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2188, authored by @mccanne,
has been merged.

simplify ast unpacker

This commit simplifies the ast unpacker by adding a generic package
for umarshaling interface types by keying on a configurable field
in the json.  The value of said key indicates which Go struct type
to unmarshal into via its type name (though this can be overriden
as evidence in our current ast where the reflector AddAs() method
is used).  Also, this requires that the Op field be filled in for
anything marshaled, which is good habit anyway because forthcoming code
that transforms ASTs and sends the transformed AST over the network
to workers must have all the Op fields present.