zq update through "add ast output formatter and "summarize" keyword" by mccanne
This is an auto-generated commit with a zq dependency update. The zq PR
https://github.com/brimsec/zq/pull/2217, authored by @mccanne,
has been merged.

add ast output formatter and "summarize" keyword

This command adds a new package zfmt to output ast's in debuggable
and canonical form.  There are two format variants as we envision
the Z kernel DSL to diverge from the Z syntax over time.  Currently,
this functionality is exposed only through the ast command but
a future PR will add a general output flag option to zq etc once this
approach is stable.

We also added an experimental "summarize" keyword to the Z language
as an optional prefix of an aggregation to go with the canonical
form of output.

We are merging a protoype of this functionality here to experiment
with and gain feedback.  This is a low-risk commit while we fine
tune zfmt since it is only exposed through the ast command for now.

This PR addresses brimsec/zq#2197 but does not close it as I would like to merge this
experimental version first, get feedback, then finalize brimsec/zq#2197 later.